Fixes/water temp to draw arg

+++ Il2ServerAppLateNight/Il2ServerAppLateNight/src/Main/Main.h
@@ -66,4 +66,4 @@ double GetCylinderHeadTemp(int engine);
 
 void UpdateCarbMixTemps();
 
-double GetCarbMixTemp(int engine);
+float GetCarbMixTemp(int engine);
